implement accept phase log replication and quorum commit
- Add AcceptRequest/AcceptResponse/AcceptLogEntry/ConflictHint messages
  (all integer fields as u32; single entry per log entry; member_id on response)
- Update accept.proto to match, shifting AcceptResponse field numbers for member_id
- Leader: dispatch accept requests to all followers concurrently via try_join_all;
  handle accept responses with fast conflict backtracking (short log, term mismatch
  branch A/B) and score board updates; advance committed slot when a quorum of
  match_slots (top N-1 followers + leader last_log_slot) exceeds committed_slot
- Follower: handle accept requests with three conflict scenarios; on anchor match,
  write entries to WAL and uncommitted_logs (always overwrite), then advance
  committed_slot to min(leader_commit_slot, last_written_slot) and evict committed
  entries; extend leader lease on every non-stale accept request
- WAL: make get_term and find_lowest_slot_for_term return Option; callers use .expect
- State machine: fold machine.rs into mod.rs; add extend_leader_lease_expiry_time to ctx


AI-Model: claude-sonnet-4.6

+++ spx_core/src/event.rs
@@ -49,8 +49,8 @@ impl PaxosEvent {
             PaxosEvent::PreVoteResponseReceived(r) => Some(r.term),
             PaxosEvent::VoteRequestReceived(cmd) => Some(cmd.get_request().term),
             PaxosEvent::VoteResponseReceived(r) => Some(r.term),
-            PaxosEvent::AcceptRequestReceived(cmd) => Some(cmd.get_request().term as u32),
-            PaxosEvent::AcceptResponseReceived(r) => Some(r.term as u32),
+            PaxosEvent::AcceptRequestReceived(cmd) => Some(cmd.get_request().term),
+            PaxosEvent::AcceptResponseReceived(r) => Some(r.term),
         }
     }
 }
